move convertBytes to a separate file
